Remove video.js things, embrace youtube hosting
When I first did this, the point was to show off that I could make a
page that served video or fell back to youtube (which was using flash).
Youtube doesn't use flash anymore and I would like to save ~$0.19 every
time someone decides to actually watch the videos so I am removing the
video.js files and just embedding youtube directly.